+++ commands/dbaas/postgres-v2/cluster/create.go
@@ -60,10 +60,11 @@ Required values to run command:
 		return []string{"4GB", "8GB", "16GB", "32GB", "64GB", "128GB", "240GB"}, cobra.ShellCompDirectiveNoFileComp
 	})
 	create.AddStringFlag(constants.FlagBackupLocation, constants.FlagBackupLocationShortPsql, "eu-central-4",
-		"The S3 location where the backups will be stored")
+		"The Object Storage location where the backups will be stored")
 	_ = create.Command.RegisterFlagCompletionFunc(constants.FlagBackupLocation, func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
 		return completer.BackupLocations(), cobra.ShellCompDirectiveNoFileComp
 	})
+	create.AddIntFlag(constants.FlagBackupRetentionDays, "", 0, "Configures how many days cluster backups are retained. Minimum: 1, Maximum: 365", core.RequiredFlagOption())
 	create.AddSetFlag(constants.FlagSyncModeV2, constants.FlagSyncModeShort, "ASYNCHRONOUS", []string{"ASYNCHRONOUS", "STRICTLY_SYNCHRONOUS"}, "Replication mode")
 	create.AddStringFlag(constants.FlagStorageSize, "", "20GB", "The amount of storage per instance in GB. Minimum: 10, Maximum: 4096. e.g.: --storage-size 20, --storage-size 20GB")
 	_ = create.Command.RegisterFlagCompletionFunc(constants.FlagStorageSize, func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
@@ -117,6 +118,10 @@ func PreRunClusterCreate(c *core.PreCommandConfig) error {
 	instances := viper.GetInt32(core.GetFlagName(c.NS, constants.FlagInstances))
 	if instances < 1 || instances > 5 {
 		return fmt.Errorf("--instances must be between 1 and 5 (got %d)", instances)
+	}
+	retentionDays := viper.GetInt32(core.GetFlagName(c.NS, constants.FlagBackupRetentionDays))
+	if retentionDays < 1 || retentionDays > 365 {
+		return fmt.Errorf("--backup-retention-days must be between 1 and 365 (got %d)", retentionDays)
 	}
 	if viper.IsSet(core.GetFlagName(c.NS, constants.FlagRecoveryTime)) && !viper.IsSet(core.GetFlagName(c.NS, constants.FlagBackupId)) {
 		return fmt.Errorf("--recovery-time requires --backup-id to be set")
@@ -185,8 +190,10 @@ func getCreateClusterRequest(c *core.CommandConfig) (psqlv2.ClusterCreate, error
 	input.SetInstances(instanceConfig)
 
 	backupLoc := viper.GetString(core.GetFlagName(c.NS, constants.FlagBackupLocation))
-	c.Verbose("BackupLocation: %v", backupLoc)
-	input.SetBackupLocation(backupLoc)
+	retentionDays := viper.GetInt32(core.GetFlagName(c.NS, constants.FlagBackupRetentionDays))
+	c.Verbose("Backup - Location: %v", backupLoc)
+	c.Verbose("Backup - RetentionDays: %v", retentionDays)
+	input.SetBackup(*psqlv2.NewClusterBackup(backupLoc, retentionDays))
 
 	displayName := viper.GetString(core.GetFlagName(c.NS, constants.FlagName))
 	c.Verbose("DisplayName: %v", displayName)
@@ -260,7 +267,7 @@ func getCreateClusterRequest(c *core.CommandConfig) (psqlv2.ClusterCreate, error
 	if viper.IsSet(core.GetFlagName(c.NS, constants.FlagBackupId)) ||
 		viper.IsSet(core.GetFlagName(c.NS, constants.FlagRecoveryTime)) {
 		backupId := viper.GetString(core.GetFlagName(c.NS, constants.FlagBackupId))
-		restoreFromBackup := psqlv2.NewPostgresClusterFromBackup(backupId)
+		restoreFromBackup := psqlv2.NewPostgresRestoreClusterFromBackup(backupId)
 
 		if viper.IsSet(core.GetFlagName(c.NS, constants.FlagRecoveryTime)) {
 			recoveryTargetTime, err := time.Parse(time.RFC3339, viper.GetString(core.GetFlagName(c.NS, constants.FlagRecoveryTime)))
@@ -277,7 +284,8 @@ func getCreateClusterRequest(c *core.CommandConfig) (psqlv2.ClusterCreate, error
 			c.Verbose("From Backup - BackupId: %v", backupId)
 		}
 
-		input.RestoreFromBackup = restoreFromBackup
+		restore := psqlv2.PostgresRestoreClusterFromBackupAsClusterRestoreFromBackup(restoreFromBackup)
+		input.RestoreFromBackup = &restore
 	}
 
 	inputCluster.SetProperties(input)